Add more documentation to README.md
Mostly just copy-paste from ASTInterpreter README but with a couple of changes based on what I've been able to gather in playing with it a bit (e.g. can install with `Pkg.add("ASTInterpreter2")`, `@enter` is a macro, etc.

I was really happy to find this as I was under the mistaken impression that there was basically no debugging functionality currently working for Julia 0.6 other than printf-style.