Work around some bugs in difficult cases
It's difficult stepping into inference because it uses some of the same types
used here to mark slots/ssas, and as a consequence `lookup_var` and others can
get confused. There also seem to be some issue wrt generated functions and/or
keyword args. So these changes just prevent recursion in known-bad cases.